Everything refactored, now comes the painstaking part
